chore: fix barrel exports
